refactor(layout): extract SiteChrome from the root layout
Header, <main> and Footer move into one server component so a sibling route
group can render full-bleed pages without them. No visual change yet.

Refs #364
