+++ src/app/layout.tsx
@@ -1,5 +1,3 @@
-import Footer from '@/components/Footer';
-import Header from '@/components/Header';
 import { SITE_URL } from '@/lib/site';
 import { Analytics } from '@vercel/analytics/react';
 import { SpeedInsights } from '@vercel/speed-insights/next';
@@ -60,18 +58,6 @@ export default function RootLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const styles = {
-    main: [
-      'relative',
-      'flex',
-      'min-h-[85vh]',
-      'flex-col',
-      'items-center',
-      'justify-center',
-      'p-24',
-    ].join(' '),
-  };
-
   const jsonLd = {
     '@context': 'https://schema.org',
     '@graph': [
@@ -111,10 +97,8 @@ export default function RootLayout({
           type="application/ld+json"
           dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
         />
-        <Header />
-        <main className={styles.main}>{children}</main>
+        {children}
         <Analytics />
-        <Footer />
         <SpeedInsights />
       </body>
     </html>
